# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления байками",
        "title": "Bike Microservice API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/components": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Добавление компонента к байку",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "components"
                ],
                "summary": "Создать компонент",
                "parameters": [
                    {
                        "description": "Данные компонента",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.ComponentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Компонент создан",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_adapter_handler_http.successResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/webike_bike_microservice_nikita_internal_core_domain.Component"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    }
                }
            }
        },
        "/components/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получение информации о компоненте по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "components"
                ],
                "summary": "Получить компонент",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID компонента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Компонент найден",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_adapter_handler_http.successResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/webike_bike_microservice_nikita_internal_core_domain.Component"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Компонент не найден",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновление данных компонента",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "components"
                ],
                "summary": "Обновить компонент",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID компонента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.UpdateComponent"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Компонент обновлен",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_adapter_handler_http.successResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/webike_bike_microservice_nikita_internal_core_domain.Component"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаление компонента",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "components"
                ],
                "summary": "Удалить компонент",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID компонента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Компонент удален",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.successResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "internal_adapter_handler_http.ComponentRequest": {
            "type": "object",
            "required": [
                "bike_id",
                "installed_mileage",
                "max_mileage",
                "name"
            ],
            "properties": {
                "bike_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "brand": {
                    "type": "string",
                    "example": "Shimano"
                },
                "installed_mileage": {
                    "type": "integer",
                    "example": 1000
                },
                "max_mileage": {
                    "type": "integer",
                    "example": 5000
                },
                "model": {
                    "type": "string",
                    "example": "Deore XT"
                },
                "name": {
                    "type": "string",
                    "example": "handlebars"
                }
            }
        },
...
        "internal_adapter_handler_http.UpdateComponent": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string",
                    "example": "Shimano"
                },
                "installed_mileage": {
                    "type": "integer",
                    "example": 1000
                },
                "max_mileage": {
                    "type": "integer",
                    "example": 5000
                },
                "model": {
                    "type": "string",
                    "example": "XT"
                },
                "name": {
                    "type": "string",
                    "example": "handlebars"
                }
            }
        },
...
        "internal_adapter_handler_http.errorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Error"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "internal_adapter_handler_http.successResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "Success message"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "webike_bike_microservice_nikita_internal_core_domain.Component": {
            "type": "object",
            "required": [
                "bike_id",
                "installed_at",
                "max_mileage",
                "name"
            ],
            "properties": {
                "bike_id": {
                    "type": "string"
                },
                "brand": {
                    "type": "string",
                    "maxLength": 100
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "installed_at": {
                    "type": "string"
                },
                "installed_mileage": {
                    "type": "integer",
                    "minimum": 0
                },
                "max_mileage": {
                    "type": "integer",
                    "maximum": 1000000,
                    "minimum": 1
                },
                "model": {
                    "type": "string",
                    "maxLength": 100
                },
                "name": {
                    "enum": [
                        "handlebars",
                        "frame",
                        "wheels"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/webike_bike_microservice_nikita_internal_core_domain.ComponentName"
                        }
                    ]
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "webike_bike_microservice_nikita_internal_core_domain.ComponentName": {
            "type": "string",
            "enum": [
                "handlebars",
                "frame",
                "wheels"
            ],
            "x-enum-varnames": [
                "Handlebars",
                "Frame",
                "Wheels"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}